backend/requirements.txt
openai
openai-agents
fastapi
uvicorn[standard]
playwright
beautifulsoup4
pandas
numpy
jinja2
python-dotenv